+++ packages/forms/src/DepositFunds.tsx
@@ -34,7 +34,7 @@ import { useQueryClient } from '@tanstack/react-query';
 import _ from 'lodash';
 import { useEffect, useMemo, useState } from 'react';
 import { useForm } from 'react-hook-form';
-import { formatEther, formatUnits, Hex, parseUnits } from 'viem';
+import { formatUnits, Hex, parseUnits } from 'viem';
 import { useAccount, useBalance, useChainId } from 'wagmi';
 
 export function DepositFunds({
@@ -278,7 +278,7 @@ export function DepositFunds({
             </Alert>
           )}
 
-          {displayBalance && displayBalance < formatEther(amount) && (
+          {!hasAmount && (
             <Alert status="error" borderRadius="md">
               <AlertIcon />
               <AlertTitle fontSize="sm">