+++ packages/constants/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @smartinvoicexyz/constants
 
+## 0.1.23
+
+### Patch Changes
+
+- fixed subgraph
+
+## 0.1.22
+
+### Patch Changes
+
+- added support for latest subgraph + fixed useEscrowZap
+
+## 0.1.21
+
+### Patch Changes
+
+- fixed useDebounce export
+
+## 0.1.20
+
+### Patch Changes
+
+- added support for client receiver
+
 ## 0.1.19
 
 ### Patch Changes