+++ packages/graphql/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @smartinvoicexyz/graphql
 
+## 0.1.23
+
+### Patch Changes
+
+- fixed subgraph
+- Updated dependencies
+  - @smartinvoicexyz/constants@0.1.23
+  - @smartinvoicexyz/shared@0.1.23
+
+## 0.1.22
+
+### Patch Changes
+
+- added support for latest subgraph + fixed useEscrowZap
+- Updated dependencies
+  - @smartinvoicexyz/constants@0.1.22
+  - @smartinvoicexyz/shared@0.1.22
+
+## 0.1.21
+
+### Patch Changes
+
+- fixed useDebounce export
+- Updated dependencies
+  - @smartinvoicexyz/constants@0.1.21
+  - @smartinvoicexyz/shared@0.1.21
+
+## 0.1.20
+
+### Patch Changes
+
+- added support for client receiver
+- Updated dependencies
+  - @smartinvoicexyz/constants@0.1.20
+  - @smartinvoicexyz/shared@0.1.20
+
 ## 0.1.19
 
 ### Patch Changes